SPT-1969: Renamed ConfirmDetailsLambda
* Renamed ConfirmDetailsLambda to ConfirmDetailsHandler, which now
  follows our preferred naming strategy.
* Moved ConfirmDetailsSubmissionHandler to it's own sub folder.
