+++ infrastructure/identity-reuse-service/template.yaml
@@ -1151,7 +1151,43 @@ Resources:
               - sts:AssumeRole
       Path: "/"
 
-  ConfirmDetailsRole:
+  GetConfirmDetailsHandler:
+    Type: AWS::Serverless::Function
+    Condition: IsNotProduction
+    Properties:
+      FunctionName: !Sub "${AWS::StackName}-GetConfirmDetailsHandler"
+      Handler: get-confirm-details-handler.lambdaHandler
+      Role: !GetAtt GetConfirmDetailsHandlerRole.Arn
+      Environment:
+        Variables:
+          NODE_OPTIONS: "--enable-source-maps"
+      Layers:
+        - !Ref GovUKFrontEndLayer
+        - !If
+          - DynaTraceEnabled
+          - !Ref DynatraceNodeLayerArn
+          - !Ref AWS::NoValue
+    Metadata:
+      BuildMethod: esbuild
+      BuildProperties:
+        UseNpmCi: True
+        EntryPoints:
+          - src/handlers/get-confirm-details-handler/get-confirm-details-handler.ts
+        Sourcemap: true
+        SourcesContent: false
+        Target: "es2024"
+        Format: esm
+        Platform: node
+        OutExtension:
+          - .js=.mjs
+        Banner:
+          - js=import { createRequire } from 'module'; const require = createRequire(import.meta.url);
+        Loader:
+          - .njk=file
+        External:
+          - govuk-frontend
+
+  GetConfirmDetailsHandlerRole:
     # checkov:skip=CKV_AWS_111: Ensure IAM policies does not allow write access without constraints
     Type: AWS::IAM::Role
     Condition: IsNotProduction
@@ -1172,68 +1208,12 @@ Resources:
         - arn:aws:iam::aws:policy/AmazonS3ReadOnlyAccess
         - arn:aws:iam::aws:policy/service-role/AWSLambdaVPCAccessExecutionRole
 
-  ConfirmDetailsLambda:
-    Type: AWS::Serverless::Function
-    Condition: IsNotProduction
-    Properties:
-      FunctionName: !Sub "${AWS::StackName}-ConfirmDetailsLambda"
-      Handler: confirm-details-handler.lambdaHandler
-      Role: !GetAtt ConfirmDetailsRole.Arn
-      Environment:
-        Variables:
-          NODE_OPTIONS: "--enable-source-maps"
-      Layers:
-        - !Ref GovUKFrontEndLayer
-        - !If
-          - DynaTraceEnabled
-          - !Ref DynatraceNodeLayerArn
-          - !Ref AWS::NoValue
-    Metadata:
-      BuildMethod: esbuild
-      BuildProperties:
-        UseNpmCi: True
-        EntryPoints:
-          - src/handlers/confirm-details/confirm-details-handler.ts
-        Sourcemap: true
-        SourcesContent: false
-        Target: "es2024"
-        Format: esm
-        Platform: node
-        OutExtension:
-          - .js=.mjs
-        Banner:
-          - js=import { createRequire } from 'module'; const require = createRequire(import.meta.url);
-        Loader:
-          - .njk=file
-        External:
-          - govuk-frontend
-
-  ConfirmDetailsSubmissionLambda:
-    Type: AWS::Serverless::Function
-    Condition: IsNotProduction
-    Properties:
-      FunctionName: !Sub "${AWS::StackName}-ConfirmDetailsSubmissionLambda"
-      Handler: confirm-details-submission-handler.lambdaHandler
-      Role: !GetAtt ConfirmDetailsRole.Arn
-      Environment:
-        Variables:
-          NODE_OPTIONS: "--enable-source-maps"
-    Metadata:
-      BuildMethod: esbuild
-      BuildProperties:
-        UseNpmCi: True
-        EntryPoints:
-          - src/handlers/confirm-details/confirm-details-submission-handler.ts
-        Target: es2022
-        Minify: true
-        Sourcemap: true
-
-  ConfirmDetailsLambdaRestApiRole:
+  GetConfirmDetailsHandlerRestApiRole:
     Type: AWS::IAM::Role
     Condition: IsNotProduction
     Properties:
-      RoleName: !Sub "${AWS::StackName}-ConfirmDetailsLambdaRestApiRole"
-      Description: !Sub "Roles for the ConfirmDetailsLambda"
+      RoleName: !Sub "${AWS::StackName}-GetConfirmDetailsHandlerRestApiRole"
+      Description: !Sub "Roles for the GetConfirmDetailsHandler"
       PermissionsBoundary: !If [ UsePermissionsBoundary, !Ref PermissionsBoundary, !Ref AWS::NoValue ]
       AssumeRolePolicyDocument:
         Version: 2012-10-17
@@ -1255,11 +1235,11 @@ Resources:
         - Key: Source
           Value: !Ref SourceTagValue
 
-  ConfirmDetailsLambdaRestApiPolicy:
+  GetConfirmDetailsHandlerRestApiPolicy:
     Type: AWS::IAM::Policy
     Condition: IsNotProduction
     Properties:
-      PolicyName: !Sub "${AWS::StackName}-ConfirmDetailsLambdaRestApiPolicy"
+      PolicyName: !Sub "${AWS::StackName}-GetConfirmDetailsHandlerRestApiPolicy"
       PolicyDocument:
         Version: 2012-10-17
         Statement:
@@ -1267,16 +1247,15 @@ Resources:
             Action:
               - "lambda:InvokeFunction"
             Resource:
-              - !Sub "arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${AWS::StackName}-ConfirmDetailsLambda:*"
-              - !Sub "arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${AWS::StackName}-ConfirmDetailsSubmissionLambda:*"
+              - !Sub "arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${AWS::StackName}-GetConfirmDetailsHandler:*"
       Roles:
-        - !Ref ConfirmDetailsLambdaRestApiRole
+        - !Ref GetConfirmDetailsHandlerRestApiRole
 
-  ConfirmDetailsFunctionLogGroup:
+  GetConfirmDetailsHandlerLogGroup:
     Type: AWS::Logs::LogGroup
     Condition: IsNotProduction
     Properties:
-      LogGroupName: !Sub "/aws/lambda/${ConfirmDetailsLambda}"
+      LogGroupName: !Sub "/aws/lambda/${AWS::StackName}-GetConfirmDetailsHandler"
       RetentionInDays: 30
       KmsKeyId: !GetAtt CloudWatchEncryptionKey.Arn
       Tags:
@@ -1291,7 +1270,48 @@ Resources:
         - Key: Source
           Value: !Ref SourceTagValue
         - Key: Name
-          Value: !Sub "/aws/lambda/${ConfirmDetailsLambda}"
+          Value: !Sub "/aws/lambda/${AWS::StackName}-GetConfirmDetailsHandler"
+
+  ConfirmDetailsSubmissionLambda:
+    Type: AWS::Serverless::Function
+    Condition: IsNotProduction
+    Properties:
+      FunctionName: !Sub "${AWS::StackName}-ConfirmDetailsSubmissionLambda"
+      Handler: confirm-details-submission-handler.lambdaHandler
+      Role: !GetAtt ConfirmDetailsSubmissionLambdaRole.Arn
+      Environment:
+        Variables:
+          NODE_OPTIONS: "--enable-source-maps"
+    Metadata:
+      BuildMethod: esbuild
+      BuildProperties:
+        UseNpmCi: True
+        EntryPoints:
+          - src/handlers/confirm-details-submission-handler/confirm-details-submission-handler.ts
+        Target: es2022
+        Minify: true
+        Sourcemap: true
+
+  ConfirmDetailsSubmissionLambdaRole:
+    # checkov:skip=CKV_AWS_111: Ensure IAM policies does not allow write access without constraints
+    Type: AWS::IAM::Role
+    Condition: IsNotProduction
+    Properties:
+      PermissionsBoundary: !If
+        - UsePermissionsBoundary
+        - !Ref PermissionsBoundary
+        - !Ref AWS::NoValue
+      AssumeRolePolicyDocument:
+        Version: 2012-10-17
+        Statement:
+          - Effect: Allow
+            Principal:
+              Service: lambda.amazonaws.com
+            Action:
+              - "sts:AssumeRole"
+      ManagedPolicyArns:
+        - arn:aws:iam::aws:policy/AmazonS3ReadOnlyAccess
+        - arn:aws:iam::aws:policy/service-role/AWSLambdaVPCAccessExecutionRole
 
   ConfirmDetailsSubmissionFunctionLogGroup:
     Type: AWS::Logs::LogGroup
